+++ src/main/java/insane96mcp/iguanatweaksreborn/module/sleeprespawn/utils/EnergyBoostItem.java
@@ -1,90 +1,80 @@
 package insane96mcp.iguanatweaksreborn.module.sleeprespawn.utils;
 
-import insane96mcp.iguanatweaksreborn.utils.LogHelper;
+import com.google.gson.*;
+import com.google.gson.annotations.JsonAdapter;
+import com.google.gson.reflect.TypeToken;
 import insane96mcp.insanelib.util.IdTagMatcher;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraftforge.registries.ForgeRegistries;
-import org.apache.commons.lang3.math.NumberUtils;
+import net.minecraft.util.GsonHelper;
 
-import javax.annotation.Nullable;
 import java.util.ArrayList;
-import java.util.List;
 
+@JsonAdapter(EnergyBoostItem.Serializer.class)
 public class EnergyBoostItem extends IdTagMatcher {
     public int duration;
     public int amplifier;
 
-    public EnergyBoostItem(Type type, ResourceLocation location) {
+    public EnergyBoostItem(Type type, String location) {
         super(type, location);
         this.duration = 0;
         this.amplifier = 0;
     }
 
-    public EnergyBoostItem(Type type, ResourceLocation location, int duration, int amplifier) {
+    public EnergyBoostItem(Type type, String location, int duration, int amplifier) {
         super(type, location);
         this.duration = duration;
         this.amplifier = amplifier;
     }
 
-    @Nullable
-    public static EnergyBoostItem parseLine(String line) {
-        String[] split = line.split(",");
-        if (split.length != 1 && split.length != 3) {
-            LogHelper.warn("Invalid line \"%s\" for Energy Boost Item", line);
-            return null;
-        }
-        Type type;
-        ResourceLocation resourceLocation;
-        if (split[0].startsWith("#")) {
-            String replaced = split[0].replace("#", "");
-            ResourceLocation tag = ResourceLocation.tryParse(replaced);
-            if (tag == null) {
-                LogHelper.warn("%s tag for Energy Boost Item is not valid", replaced);
-                return null;
-            }
-            type = Type.TAG;
-            resourceLocation = tag;
-        }
-        else {
-            ResourceLocation item = ResourceLocation.tryParse(split[0]);
-            if (item == null) {
-                LogHelper.warn("%s item for Energy Boost Item is not valid", split[0]);
-                return null;
-            }
-            if (!ForgeRegistries.ITEMS.containsKey(item)) {
-                LogHelper.warn(String.format("%s item for Energy Boost Item seems to not exist", split[0]));
-                return null;
-            }
-            type = Type.ID;
-            resourceLocation = item;
-        }
-        if (split.length == 1) {
-            return new EnergyBoostItem(type, resourceLocation);
-        }
-        else {
-            if (!NumberUtils.isParsable(split[1])) {
-                LogHelper.warn(String.format("Invalid duration \"%s\" for Energy Boost Item", line));
-                return null;
-            }
-            int duration = Integer.parseInt(split[1]);
+    public static final java.lang.reflect.Type LIST_TYPE = new TypeToken<ArrayList<EnergyBoostItem>>(){}.getType();
+    public static class Serializer implements JsonDeserializer<EnergyBoostItem>, JsonSerializer<EnergyBoostItem> {
+        @Override
+        public EnergyBoostItem deserialize(JsonElement json, java.lang.reflect.Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
+            String id = GsonHelper.getAsString(json.getAsJsonObject(), "id", "");
+            String tag = GsonHelper.getAsString(json.getAsJsonObject(), "tag", "");
 
-            if (!NumberUtils.isParsable(split[2])) {
-                LogHelper.warn(String.format("Invalid amplifier \"%s\" for Energy Boost Item", line));
-                return null;
+            if (!id.equals("") && !ResourceLocation.isValidResourceLocation(id)) {
+                throw new JsonParseException("Invalid id: %s".formatted(id));
             }
-            int amplifier = Integer.parseInt(split[2]);
-            return new EnergyBoostItem(type, resourceLocation, duration, amplifier);
-        }
-    }
+            if (!tag.equals("") && !ResourceLocation.isValidResourceLocation(id)) {
+                throw new JsonParseException("Invalid tag: %s".formatted(tag));
+            }
 
-    public static ArrayList<EnergyBoostItem> parseStringList(List<? extends String> list) {
-        ArrayList<EnergyBoostItem> stackSizes = new ArrayList<>();
-        for (String line : list) {
-            EnergyBoostItem energyBoostItem = EnergyBoostItem.parseLine(line);
-            if (energyBoostItem != null)
-                stackSizes.add(energyBoostItem);
+            EnergyBoostItem energyBoostItem;
+            if (!id.equals("") && !tag.equals("")){
+                throw new JsonParseException("Invalid object containing both tag (%s) and id (%s)".formatted(tag, id));
+            }
+            else if (!id.equals("")) {
+                energyBoostItem = new EnergyBoostItem(Type.ID, id);
+            }
+            else if (!tag.equals("")){
+                energyBoostItem = new EnergyBoostItem(Type.TAG, tag);
+            }
+            else {
+                throw new JsonParseException("Invalid object missing either tag and id");
+            }
+
+            energyBoostItem.duration = GsonHelper.getAsInt(json.getAsJsonObject(), "duration", 0);
+            energyBoostItem.amplifier = GsonHelper.getAsInt(json.getAsJsonObject(), "amplifier", 0);
+
+            return energyBoostItem;
         }
-        return stackSizes;
+
+        @Override
+        public JsonElement serialize(EnergyBoostItem src, java.lang.reflect.Type typeOfSrc, JsonSerializationContext context) {
+            JsonObject jsonObject = new JsonObject();
+            if (src.type == Type.ID) {
+                jsonObject.addProperty("id", src.location.toString());
+            }
+            else if (src.type == Type.TAG) {
+                jsonObject.addProperty("tag", src.location.toString());
+            }
+            jsonObject.addProperty("duration", src.duration);
+            if (src.amplifier > 0)
+                jsonObject.addProperty("amplifier", src.amplifier);
+
+            return jsonObject;
+        }
     }
 
 }